Fix decimal when prev operand contains one
If previous operand contains a decimal point, then the check for an
existing decimal point in the displayValue will disable the button when
it should be available for building the new operand.

For this reason, I moved the operand is null check to be the first if
statement and added an else to the displayValue.includes check.